Deleting a Reservation

+++ app/_lib/actions.js
@@ -30,6 +30,12 @@ export async function deleteReservation(bookingId) {
   const session = await auth();
   if (!session) throw new Error("You must be logged in");
 
+  const guestBookings = await getBookings(SquareAsteriskIcon.user.guestId);
+  const guestBookingIds = guestBookings.map((booking) => booking.id);
+
+  if (!guestBookingIds.includes(bookingId))
+    throw new Error("You are not allowed to delte this booking");
+
   const { error } = await supabase
     .from("bookings")
     .delete()
